Break at the end
Assuming that the normal break at the end of the loop is removed, the
pagination should eventually terminate, and reach a steady state where
the query returns 0 results (because there are no entity IDs above the
entity_id_offset, and no page IDs above the page_id_offset for the same
entity ID). In that case, break the loop (finish the script) instead of
repeating the same query over and over again (in a tight loop, since
this case would never reach the time.sleep()).